requirements.txt
Flask
azure-cognitiveservices-speech
azure-ai-textanalytics
pydub
soundfile
gunicorn
